+++ Assets/Scripts/Game/NPCs/DialogueGenerator/DialogueQuestCheckpoint.cs
@@ -1,7 +1,5 @@
-using System.Text;
-using Game.NarrativeGenerator.Quests;
+using Overlord.NarrativeGenerator.Quests;
 using Game.NarrativeGenerator.Quests.QuestGrammarTerminals;
-using MyBox;
 using UnityEngine;
 
 namespace Game.NPCs
