+++ Assets/Scripts/Game/NPCs/DialogueGenerator/DialogueQuestCloser.cs
@@ -1,7 +1,6 @@
 using System.Text;
-using Game.NarrativeGenerator.Quests;
+using Overlord.NarrativeGenerator.Quests;
 using Game.NarrativeGenerator.Quests.QuestGrammarTerminals;
-using MyBox;
 using UnityEngine;
 
 namespace Game.NPCs
