+++ Assets/Scripts/Game/NPCs/DialogueGenerator/QuestDialogue.cs
@@ -1,8 +1,5 @@
-using System.Linq;
 using System.Text;
-using Game.NarrativeGenerator.Quests;
-using Game.NarrativeGenerator.Quests.QuestGrammarTerminals;
-using MyBox;
+using Overlord.NarrativeGenerator.Quests;
 using UnityEngine;
 
 namespace Game.NPCs
