+++ Assets/Scripts/Game/NarrativeGenerator/DungeonRelatedNarrative/QuestDungeonsParameters.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using static Util.Enums;
+using Overlord.NarrativeGenerator.Quests;
 
 namespace Game.NarrativeGenerator
 {
