+++ Assets/Scripts/Game/NarrativeGenerator/Quests/ProfileQuests/Immersion/GiveQuestSo.cs
@@ -3,6 +3,7 @@ using Util;
 using System.Collections.Generic;
 using System;
 using Game.NPCs;
+using Overlord.NarrativeGenerator.Quests;
 
 namespace Game.NarrativeGenerator.Quests.QuestGrammarTerminals
 {
