+++ Assets/Scripts/Game/NarrativeGenerator/Quests/QuestSo.cs
@@ -5,12 +5,11 @@ using UnityEngine;
 using Util;
 using Game.NPCs;
 
-
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
 
-namespace Game.NarrativeGenerator.Quests
+namespace Overlord.NarrativeGenerator.Quests
 {
 
     [CreateAssetMenu(fileName = "QuestSo", menuName = "Overlord-Project/QuestSo", order = 0)]
