+++ Assets/Scripts/Overlord/NarrativeGenerator/QuestGenerationAlgorithm.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c70e1589961658499aa6e18abd1fbe6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
